drishti-mother-child/src/main/java/org/ei/drishti/service/formSubmission/handler/FPReferralFollowupHandler.java
package org.ei.drishti.service.formSubmission.handler;

import org.ei.drishti.form.domain.FormSubmission;
import org.ei.drishti.service.ECService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Component;

@Component
public class FPReferralFollowupHandler implements FormSubmissionHandler {
    private ECService ecService;

    @Autowired
    public FPReferralFollowupHandler(ECService ecService) {
        this.ecService = ecService;
    }

    @Override
    public void handle(FormSubmission submission) {
        ecService.reportReferralFollowup(submission);
    }
}